Enable all File-like objects to be used for uploading
Hard-coding the `File` class as a condition for uploading a file prevents other legitimate IO-compatible stream objects (e.g. `StringIO`) from being used as input.  This can cause cryptic errors to be returned to the end user from the Gemfury server about `multipart/form-data` not being used, when in fact the Faraday request middleware has failed to recognize a file being supplied.

This patch loosens the criteria to simply testing for some of the methods that Faraday will need to process the input stream: 
https://github.com/lostisland/faraday/blob/master/lib/faraday/file_part.rb

Experimentally, although ["rewind", "read", "close", "length", and "ensure_open_and_readable"] are referred to in Faraday's `CompositeReadIO`, passing an object with ["rewind", "read", "close"] implemented seems to produce the expected behavior